+++ flow-typed/globals.js
@@ -1,0 +1,11 @@
+/**
+ * Makes flow aware of global variables defined by the webpack define plugin.
+ *
+ * @flow
+ */
+
+// eslint-disable-next-line no-unused-vars
+declare var __ACTION__: Object;
+
+// eslint-disable-next-line no-unused-vars
+declare var __RELEASE_FLAGS__: Object;