+++ src/universal/components/RetroRoot/RetroRoot.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+export default () => (
+  <div
+    style={{
+      height: '100vh',
+      display: 'flex',
+      alignItems: 'center',
+      backgroundColor: 'rgba(256, 0, 0, 0.5)'
+    }}
+  >
+    <div style={{fontSize: '4rem', margin: 'auto'}}>~😇~</div>
+  </div>
+);