+++ src/universal/modules/outcomeCard/components/OutcomeCardFooter/OutcomeCardFooter.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import React, {Component} from 'react';
 import {createFragmentContainer} from 'react-relay';
 import withAtmosphere from 'universal/decorators/withAtmosphere/withAtmosphere';
-import AsyncMenuContainer from 'universal/modules/menu/containers/AsyncMenu/AsyncMenu';
 import OutcomeCardMessage from 'universal/modules/outcomeCard/components/OutcomeCardMessage/OutcomeCardMessage';
 import UpdateTaskMutation from 'universal/mutations/UpdateTaskMutation';
 import textOverflow from 'universal/styles/helpers/textOverflow';
@@ -19,14 +18,42 @@ import avatarUser from 'universal/styles/theme/images/avatar-user.svg';
 import Loadable from 'react-loadable';
 import LoadableMenu from 'universal/components/LoadableMenu';
 import LoadableLoading from 'universal/components/LoadableLoading';
-
-const fetchGitHubRepos = () => System.import('universal/containers/GitHubReposMenuRoot/GitHubReposMenuRoot');
-const fetchStatusMenu = () => System.import('universal/modules/outcomeCard/components/OutcomeCardStatusMenu/OutcomeCardStatusMenu');
+import PlainButton from 'universal/components/PlainButton/PlainButton';
 
 const LoadableAssignMenu = Loadable({
   loader: () => System.import(
     /* webpackChunkName: 'OutcomeCardAssignMenuRoot' */
     'universal/modules/outcomeCard/components/OutcomeCardAssignMenuRoot'
+  ),
+  loading: (props) => <LoadableLoading {...props} height={DEFAULT_MENU_HEIGHT} width={DEFAULT_MENU_WIDTH} />,
+  delay: HUMAN_ADDICTION_THRESH,
+  timeout: MAX_WAIT_TIME
+});
+
+const LoadableAssignTeamMenu = Loadable({
+  loader: () => System.import(
+    /* webpackChunkName: 'OutcomeCardAssignMenuRoot' */
+    'universal/modules/outcomeCard/components/OutcomeCardAssignTeamMenuRoot'
+  ),
+  loading: (props) => <LoadableLoading {...props} height={DEFAULT_MENU_HEIGHT} width={DEFAULT_MENU_WIDTH} />,
+  delay: HUMAN_ADDICTION_THRESH,
+  timeout: MAX_WAIT_TIME
+});
+
+const LoadableStatusMenu = Loadable({
+  loader: () => System.import(
+    /* webpackChunkName: 'OutcomeCardStatusMenu' */
+    'universal/modules/outcomeCard/components/OutcomeCardStatusMenu/OutcomeCardStatusMenu'
+  ),
+  loading: (props) => <LoadableLoading {...props} height={DEFAULT_MENU_HEIGHT} width={DEFAULT_MENU_WIDTH} />,
+  delay: HUMAN_ADDICTION_THRESH,
+  timeout: MAX_WAIT_TIME
+});
+
+const LoadableGitHubMenu = Loadable({
+  loader: () => System.import(
+    /* webpackChunkName: 'GitHubReposMenuRoot' */
+    'universal/containers/GitHubReposMenuRoot/GitHubReposMenuRoot'
   ),
   loading: (props) => <LoadableLoading {...props} height={DEFAULT_MENU_HEIGHT} width={DEFAULT_MENU_WIDTH} />,
   delay: HUMAN_ADDICTION_THRESH,
@@ -104,13 +131,17 @@ class OutcomeCardFooter extends Component {
     const {error} = this.state;
     const ownerAvatarOrTeamName = (
       showTeam ?
-        <div className={css(styles.teamNameLabel)}>{teamName}</div> :
+        (<PlainButton
+          aria-label="Assign this task to another team"
+          onClick={this.selectAllQuestion}
+        >
+          {teamName}
+        </PlainButton>) :
         (<button
+          aria-label="Assign this task to a teammate"
           className={css(styles.avatarButton)}
-          tabIndex={service && '-1'}
           title={`Assigned to ${assignee.preferredName}`}
           type="button"
-          ref={(c) => { this.assignRef = c; }}
         >
           <div className={avatarStyles}>
             <img
@@ -127,38 +158,54 @@ class OutcomeCardFooter extends Component {
         </button>)
     );
 
+    const canAssign = !service && !isArchived;
     return (
       <div className={css(styles.footerAndMessage)}>
         <div className={css(styles.footer)}>
           <div className={css(styles.avatarBlock)}>
-            {service || showTeam || isArchived ?
-              ownerAvatarOrTeamName :
-              <LoadableMenu
-                LoadableComponent={LoadableAssignMenu}
-                maxWidth={350}
-                maxHeight={225}
-                originAnchor={assignOriginAnchor}
-                queryVars={{
-                  area,
-                  assignRef: this.assignRef,
-                  task,
-                  teamId
-                }}
-                targetAnchor={assignTargetAnchor}
-                toggle={ownerAvatarOrTeamName}
-                onOpen={toggleMenuState}
-                onClose={toggleMenuState}
-              />
+            {!canAssign && ownerAvatarOrTeamName}
+            {canAssign && showTeam &&
+            <LoadableMenu
+              LoadableComponent={LoadableAssignTeamMenu}
+              maxWidth={350}
+              maxHeight={225}
+              originAnchor={assignOriginAnchor}
+              queryVars={{
+                area,
+                task
+              }}
+              targetAnchor={assignTargetAnchor}
+              toggle={ownerAvatarOrTeamName}
+              onOpen={toggleMenuState}
+              onClose={toggleMenuState}
+            />
+            }
+            {canAssign && !showTeam &&
+            <LoadableMenu
+              LoadableComponent={LoadableAssignMenu}
+              maxWidth={350}
+              maxHeight={225}
+              originAnchor={assignOriginAnchor}
+              queryVars={{
+                area,
+                task,
+                teamId
+              }}
+              targetAnchor={assignTargetAnchor}
+              toggle={ownerAvatarOrTeamName}
+              onOpen={toggleMenuState}
+              onClose={toggleMenuState}
+            />
             }
           </div>
           <div className={buttonBlockStyles}>
             {isArchived ?
               <OutcomeCardFooterButton onClick={this.removeContentTag('archived')} icon="reply" /> :
-              <div>
+              <React.Fragment>
                 {/* buttonSpacer helps truncated names (…) be consistent */}
                 {!service ?
-                  <AsyncMenuContainer
-                    fetchMenu={fetchGitHubRepos}
+                  <LoadableMenu
+                    LoadableComponent={LoadableGitHubMenu}
                     maxWidth={350}
                     maxHeight={225}
                     originAnchor={originAnchor}
@@ -166,18 +213,20 @@ class OutcomeCardFooter extends Component {
                       area,
                       handleAddTask,
                       taskId,
+                      teamId,
                       setError: this.setError,
                       clearError: this.clearError
                     }}
                     targetAnchor={targetAnchor}
                     toggle={<OutcomeCardFooterButton icon="github" />}
-                    toggleMenuState={toggleMenuState}
+                    onOpen={toggleMenuState}
+                    onClose={toggleMenuState}
                   /> :
                   <div className={css(styles.buttonSpacer)} />
                 }
-                <AsyncMenuContainer
-                  fetchMenu={fetchStatusMenu}
-                  maxWidth={200}
+                <LoadableMenu
+                  LoadableComponent={LoadableStatusMenu}
+                  maxWidth={350}
                   maxHeight={225}
                   originAnchor={originAnchor}
                   queryVars={{
@@ -190,9 +239,10 @@ class OutcomeCardFooter extends Component {
                   }}
                   targetAnchor={targetAnchor}
                   toggle={<OutcomeCardFooterButton icon="ellipsis-v" />}
-                  toggleMenuState={toggleMenuState}
+                  onOpen={toggleMenuState}
+                  onClose={toggleMenuState}
                 />
-              </div>
+              </React.Fragment>
             }
           </div>
         </div>
@@ -349,6 +399,7 @@ export default createFragmentContainer(
         teamId: id
         teamName: name
       }
+      ...OutcomeCardAssignTeamMenu_task
       ...OutcomeCardAssignMenu_task
       ...OutcomeCardStatusMenu_task
     }`