+++ src/universal/releaseFlags.js
@@ -18,4 +18,9 @@
  * @flow
  */
 
-export default {};
+export default {
+  // The single-page embedded signin
+  newSignIn: false,
+  // The retrospective meeting
+  retro: false
+};