+++ src/universal/utils/constants.js
@@ -209,8 +209,7 @@ export const TEAM_MEMBER = 'teamMember';
 // export const GQL_CONNECTION_INIT = 'connection_init'; // Client -> Server
 export const GQL_CONNECTION_ACK = 'connection_ack'; // Server -> Client
 export const GQL_CONNECTION_ERROR = 'connection_error'; // Server -> Client
-// NOTE: The keep alive message type does not follow the Apollo standard because the "standard" doesn't use native standard
-// export const GQL_CONNECTION_KEEP_ALIVE = 'ka'; // Server -> Client
+export const GQL_CONNECTION_KEEP_ALIVE = 'ka'; // Server <-> Client
 export const GQL_CONNECTION_TERMINATE = 'connection_terminate'; // Client -> Server
 export const GQL_START = 'start'; // Client -> Server
 export const GQL_DATA = 'data'; // Server -> Client
@@ -218,6 +217,8 @@ export const GQL_ERROR = 'error'; // Server -> Client
 export const GQL_COMPLETE = 'complete'; // Server -> Client
 export const GQL_STOP = 'stop'; // Client -> Server
 
+/* WebSocket keep alive interval */
+export const WS_KEEP_ALIVE = 10000;
 
 /* Relay Subscription Event Types */
 export const UPDATED = 'updated';